# Backend/openapi.yaml
openapi: 3.1.0
info:
  title: SCUD Backend API
  version: "1.0.0"
  description: |
    HTTP API системы управления доступом (SCUD).

    Контракт описывает три класса участников:

    - **Мобильное приложение пользователя** (Android) — авторизуется
      session_token'ом, выпускает ключи доступа, скачивает Bloom-фильтры для
      доставки на ридер, отправляет отчёты от ридера на сервер.
    - **Администратор / интеграция** — авторизуется API-ключом
      (`X-Api-Key`), управляет пользователями, группами, ридерами, permit'ами,
      выпущенными ключами, наблюдает доставку фильтров.
    - **Public** — health-check и пара unauthenticated auth-эндпоинтов
      (login / refresh).

    Ключевые понятия:

    - **Reader** — физический контроллер двери (ESP32). Имеет 16-байтный
      `reader_id` и Ed25519 публичный ключ.
    - **Reader group** — группа ридеров; пользователь видит только ридеры
      из доступной ему `user_group → reader_group` пары.
    - **Permit** — разрешение «пользователь P может пользоваться ридером R
      в окне [valid_from, valid_until], выпускать до n_parallel ключей с
      TTL ≤ max_token_ttl_seconds».
    - **IssuedKey** — выпущенный пользователем ключ (CARD на терминологию
      ридера). Подписан сервером, содержит `permit_id`, `device_id`, окно
      `[issued_at, expires_at]`. Отозванные ключи попадают в Bloom-фильтр
      ридера.
    - **TimeGrant** — подписанный сервером talisman, разрешающий телефону
...
      `"revoked"`).
servers:
...
security:
  - SessionToken: []
  - AdminApiKey: []
paths:
...
  /api/v1/app/permits:
    get:
      tags: [app-permits]
      summary: Permit'ы текущего пользователя
      description: |
        Возвращает активные (не отозванные, ещё валидные по времени) permit'ы.
        Для каждого: имя ридера, окно валидности, n_parallel, число активных
        ключей и есть ли уже time_grant для текущего устройства (нужно UI,
        чтобы решить, надо ли просить grant при выпуске следующего ключа).
      security:
        - SessionToken: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: {$ref: "#/components/schemas/PermitListItem"}
        "401": {$ref: "#/components/responses/Unauthorized"}
  /api/v1/app/permits/{permit_id}/keys:
    get:
      tags: [app-permits]
      summary: Все ключи permit'а
      description: |
        Возвращает все ключи permit'а (включая ключи с других устройств того
        же пользователя и неактивные/expired/revoked) — UI сам решает, что
        показывать в архиве, а что в активном списке.

        `full_key_bytes` (base64) возвращается, чтобы приложение могло
        локально кэшировать ключ для tap'а; для чужих устройств он бесполезен,
        но безвреден — без приватного ключа Ed25519 ключ не подписать.
      security:
        - SessionToken: []
      parameters:
        - {name: permit_id, in: path, required: true, schema: {type: string, format: uuid}}
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: {$ref: "#/components/schemas/IssuedKeyAppView"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {description: "not_owner"}
  /api/v1/app/permits/{permit_id}/revoke:
    post:
      tags: [app-permits]
      summary: Пользователь отзывает свой permit
      description: |
        Синхронный one-shot revoke: разрешён только когда у permit'а нет
        активных ключей (иначе **409 has_active_keys** с количеством).

        Это сознательно строже, чем admin-revoke: пользователь сам должен
        сначала отозвать все свои ключи (или дождаться истечения), а потом
        отказаться от permit'а. Двухфазной "revoking"-стадии тут нет.
      security:
        - SessionToken: []
      parameters:
        - {name: permit_id, in: path, required: true, schema: {type: string, format: uuid}}
      responses:
        "200":
          description: OK
          content: {application/json: {schema: {$ref: "#/components/schemas/Ok"}}}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {description: "not_owner"}
        "409":
          description: has_active_keys
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: object
                    properties:
                      error: {type: string, enum: ["has_active_keys"]}
                      active_count: {type: integer}

  # ===================================================================
  # APP / KEYS
  # ===================================================================
  /api/v1/app/keys/request:
    post:
      tags: [app-keys]
      summary: Выпустить новый ключ доступа
      description: |
        Главный путь выпуска access-token'а: проверяется владение permit'ом,
        окно валидности permit'а, лимит `n_parallel`, и максимально допустимый
        TTL (`max_token_ttl_seconds`). Ключ привязывается к `device_id` из
        текущей Session.

        Поле `request_grant=true` дополнительно выпускает Time-Grant —
        подписанный сервером объект, разрешающий приложению скорректировать
        часы ридера в момент tap'а, если drift > 15 сек.

        `issued_at_ts` (unix sec) позволяет выпустить future-dated ключ
        (start time picker в UI) — ключ становится валидным только начиная
        с этой метки.

        Регенерация Bloom-фильтра ридера НЕ триггерится при выпуске ключа.
        Это by-design: новый ключ добавляется в whitelist на стадии активации
        FP-fallback'а, а сам Bloom меняется только при revoke.
      security:
        - SessionToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/KeyRequest"}
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {$ref: "#/components/schemas/KeyResponse"}
        "400":
          description: |
            ttl_too_long / ttl_invalid / permit_expired / permit_not_started /
            no_active_device.
          content:
            application/json:
              schema: {$ref: "#/components/schemas/ErrorDetail"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {description: "not_owner"}
        "404": {description: "permit_not_found"}
        "409":
          description: n_parallel_exceeded
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: object
                    properties:
                      error: {type: string, enum: ["n_parallel_exceeded"]}
                      active_count: {type: integer}
                      max: {type: integer}
  /api/v1/app/keys/{key_id}/revoke-on-server:
    post:
      tags: [app-keys]
      summary: Пользовательский revoke ключа (сервер-only)
      description: |
        Помечает ключ как `revoked_by_server` (PG-trigger оставляет
        `is_active=True`, чтобы n_parallel-гейт не пропустил замену, пока
        ридер не получит новый bloom).

        Триггерит фоновую генерацию свежего фильтра для ридера → запись
        попадает в Bloom → после доставки ридер переводит ключ в
        `revoked_in_bloom` (`is_active=False`).

        Ошибка **409 not_active**, если ключ уже отозван/истёк.
      security:
        - SessionToken: []
      parameters:
        - {name: key_id, in: path, required: true, schema: {type: string, description: "Hex-ID ключа (12 байт)."}}
      responses:
        "200":
          description: OK
          content: {application/json: {schema: {$ref: "#/components/schemas/Ok"}}}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "403": {description: "not_owner"}
        "404": {description: "key_not_found"}
        "409":
          description: not_active
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: object
                    properties:
                      error: {type: string, enum: ["not_active"]}

  # ===================================================================
  # APP / COURIER
  # ===================================================================
...
  /api/v1/admin/permits:
    post:
      tags: [admin-permits]
      summary: Создать permit
      description: |
        Создаёт permit для пары (`user_id`, `reader_id`). В permit'е
        фиксируются: окно валидности, лимит параллельных ключей
        (`n_parallel`), и максимальный TTL ключа в секундах
        (`max_token_ttl_seconds`).
      security:
        - AdminApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/CreatePermitRequest"}
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  permit_id: {type: string, format: uuid}
        "401": {$ref: "#/components/responses/AdminUnauthorized"}
        "404": {description: "user_not_found / reader_not_found"}
        "422": {description: "invalid reader_id hex"}
    get:
      tags: [admin-permits]
      summary: Список permit'ов
      description: |
        Постраничный список с производным полем `status`:
        - **active** — `revoke_initiated_at == NULL`
        - **revoking** — Phase 1 admin-revoke: `revoke_initiated_at` стоит,
          но `revoked_at == NULL` (ждём пока ридер применит свежий bloom и
          ключи станут `is_active=False`).
        - **revoked** — финализирован, `revoked_at != NULL`.
      security:
        - AdminApiKey: []
      parameters:
        - {$ref: "#/components/parameters/Cursor"}
        - {$ref: "#/components/parameters/Limit"}
        - {name: user_id, in: query, required: false, schema: {type: integer}}
        - {name: reader_id, in: query, required: false, schema: {type: string, description: "hex"}}
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: {$ref: "#/components/schemas/AdminPermitListItem"}
        "401": {$ref: "#/components/responses/AdminUnauthorized"}
  /api/v1/admin/permits/{permit_id}:
    get:
      tags: [admin-permits]
      summary: Подробности permit'а + список ключей
      description: |
        Кроме самого permit'а возвращает массив всех `keys` (с `is_active`,
        `status`) и агрегат `active_keys_count` — нужно чтобы понимать,
        почему "revoking" ещё не финализировался.
      security:
        - AdminApiKey: []
      parameters:
        - {name: permit_id, in: path, required: true, schema: {type: string, format: uuid}}
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: {$ref: "#/components/schemas/AdminPermitDetail"}
        "401": {$ref: "#/components/responses/AdminUnauthorized"}
        "404": {description: "permit_not_found"}
    patch:
      tags: [admin-permits]
      summary: Изменить permit
      description: |
        Меняет атрибуты permit'а (display_name, description, n_parallel,
        max_token_ttl_seconds). Не меняет `valid_from`/`valid_until` — для
        этого надо отозвать и выпустить новый.
      security:
        - AdminApiKey: []
      parameters:
        - {name: permit_id, in: path, required: true, schema: {type: string, format: uuid}}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/PatchPermitRequest"}
      responses:
        "200": {$ref: "#/components/responses/Ok"}
        "401": {$ref: "#/components/responses/AdminUnauthorized"}
        "404": {description: "permit_not_found"}
  /api/v1/admin/permits/{permit_id}/revoke:
    post:
      tags: [admin-permits]
      summary: Admin-revoke permit'а (двухфазный)
      description: |
        **Phase 1 (синхронный):**
        - Все active-ключи permit'а помечаются `revoked_by_server`
          (`is_active` остаётся True пока ридер не применил bloom).
        - Все time-grant'ы permit'а отзываются.
        - Если active-ключей не было → permit финализируется сразу
          (`revoked_at = now`, status `"revoked"`).
        - Иначе permit входит в состояние **revoking**: ставится
          `revoke_initiated_at = now`, но `revoked_at = NULL`.
        - Для каждого затронутого ридера планируется фоновая задача
          `generate_filter` (debounced).

        **Phase 2 (асинхронный, через `process_delivery_receipt` /
        `process_fdi_blob`):**
        - Когда ключи переходят `revoked_by_server → revoked_in_bloom`
          (`is_active=False`), сервер прогоняет `finalize_pending_revokes`
          по затронутому ридеру: permit'ы с active=0 финализируются
          (`revoked_at = now`).

        Ответ возвращает текущий статус (`revoking` или `revoked`), чтобы
        админский UI мог отрисовать «отзыв в процессе».
      security:
        - AdminApiKey: []
      parameters:
        - {name: permit_id, in: path, required: true, schema: {type: string, format: uuid}}
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: {type: boolean}
                  status: {type: string, enum: [active, revoking, revoked]}
                  revoke_initiated_at: {type: string, format: date-time, nullable: true}
                  revoked_at: {type: string, format: date-time, nullable: true}
        "401": {$ref: "#/components/responses/AdminUnauthorized"}
        "404": {description: "permit_not_found"}

  # ===================================================================
  # ADMIN / KEYS
  # ===================================================================
...
components:
  securitySchemes:
    SessionToken:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: |
        `Authorization: Bearer <session_token>` — выдаётся `/auth/login` или
        `/auth/refresh`. Хранится на сервере в таблице sessions, валидируется
        каждый запрос (revoked_at IS NULL, expires_at > now).
    AdminApiKey:
      type: apiKey
      in: header
      name: X-Api-Key
      description: |
        `X-Api-Key: <plaintext>` — выдаётся `/admin/api-keys` один раз при
        создании. На сервере хранится только SHA256-хеш + `key_prefix`.
  parameters:
    Cursor:
      name: cursor
      in: query
      required: false
      schema: {type: integer, minimum: 0, default: 0}
      description: Offset для пагинации.
    Limit:
      name: limit
      in: query
      required: false
      schema: {type: integer, minimum: 1, maximum: 200, default: 50}
  responses:
    Ok:
      description: OK
      content:
        application/json:
          schema: {$ref: "#/components/schemas/Ok"}
    Unauthorized:
      description: missing_token / invalid_token / user_inactive
    AdminUnauthorized:
      description: missing_api_key / invalid_api_key / api_key_expired
  schemas:
    Ok:
      type: object
      properties:
        ok: {type: boolean, example: true}
    ErrorDetail:
      type: object
      description: |
        FastAPI стандартный формат — `{"detail": ...}`. detail может быть
        строкой (`"not_owner"`) или объектом (`{"error": "ttl_too_long",
        "max": 86400}`).
      properties:
        detail:
          oneOf:
            - {type: string}
            - {type: object}
...
    PermitListItem:
      type: object
      properties:
        permit_id: {type: string, format: uuid}
        display_name: {type: string}
        description: {type: string, nullable: true}
        reader_id: {type: string, description: "hex"}
        reader_display_name: {type: string}
        valid_from: {type: string, format: date-time}
        valid_until: {type: string, format: date-time}
        n_parallel: {type: integer}
        max_token_ttl_seconds: {type: integer}
        active_keys_count: {type: integer}
        has_grant_for_device: {type: boolean}
    IssuedKeyAppView:
      type: object
      properties:
        key_id: {type: string, description: "hex"}
        device_id: {type: string, format: uuid}
        device_label: {type: string, nullable: true}
        is_current_device: {type: boolean}
        issued_at: {type: string, format: date-time}
        expires_at: {type: string, format: date-time}
        status: {type: string, description: "active | revoked_by_server | revoked_by_reader | revoked_in_bloom | expired"}
        full_key_bytes: {type: string, description: "base64"}

    # ----- KEYS (APP) -----
    KeyRequest:
      type: object
      required: [permit_id, validity_seconds]
      properties:
        permit_id: {type: string, format: uuid}
        validity_seconds: {type: integer, minimum: 1}
        request_grant: {type: boolean, default: false}
        issued_at_ts:
          type: integer
          nullable: true
          description: Unix-секунды; None = now. Поддержка future-dated ключей.
    KeyResponse:
      type: object
      properties:
        issued_key:
          type: object
          properties:
            key_id: {type: string, description: "hex"}
            full_key_bytes: {type: string, description: "base64"}
            issued_at: {type: string, format: date-time}
            expires_at: {type: string, format: date-time}
        time_grant:
          type: object
          nullable: true
          properties:
            grant_id: {type: string, format: uuid}
            full_grant_bytes: {type: string, description: "base64"}
            expires_at: {type: string, format: date-time}

    # ----- COURIER -----
...
    CreatePermitRequest:
      type: object
      required: [user_id, reader_id, display_name, valid_from, valid_until]
      properties:
        user_id: {type: integer}
        reader_id: {type: string, description: "hex"}
        display_name: {type: string}
        description: {type: string, nullable: true}
        valid_from: {type: string, format: date-time}
        valid_until: {type: string, format: date-time}
        n_parallel: {type: integer, default: 1, minimum: 1}
        max_token_ttl_seconds: {type: integer, default: 86400, minimum: 1}
    PatchPermitRequest:
      type: object
      properties:
        display_name: {type: string}
        description: {type: string, nullable: true}
        n_parallel: {type: integer, minimum: 1}
        max_token_ttl_seconds: {type: integer, minimum: 1}
    AdminPermitListItem:
      type: object
      properties:
        permit_id: {type: string, format: uuid}
        user_id: {type: integer}
        reader_id: {type: string, description: "hex"}
        display_name: {type: string}
        valid_from: {type: string, format: date-time}
        valid_until: {type: string, format: date-time}
        n_parallel: {type: integer}
        status: {type: string, enum: [active, revoking, revoked]}
        revoke_initiated_at: {type: string, format: date-time, nullable: true}
        revoked_at: {type: string, format: date-time, nullable: true}
    AdminPermitDetail:
      allOf:
        - {$ref: "#/components/schemas/AdminPermitListItem"}
        - type: object
          properties:
            description: {type: string, nullable: true}
            max_token_ttl_seconds: {type: integer}
            active_keys_count: {type: integer}
            keys:
              type: array
              items:
                type: object
                properties:
                  key_id: {type: string, description: "hex"}
                  status: {type: string}
                  is_active: {type: boolean}
                  issued_at: {type: string, format: date-time}
                  expires_at: {type: string, format: date-time}

    # ----- KEYS (ADMIN) -----
...